+++ tests/testthat/test_prepare_results_dudi.pca.R
@@ -7,11 +7,9 @@ sup_var <- d[-(1:10), 8:9]
 sup_ind <- d[1:10, -(8:9)]
 pca <- dudi.pca(d[-(1:10), -(8:9)], scale = TRUE, scannf = FALSE, nf = 5)
 ## Supplementary individuals
-supi <- suprow(pca, sup_ind)
-pca$supi <- supi$lisup
+pca$supi <- suprow(pca, sup_ind)
 ## Supplementary variables
-supv <- supcol(pca, dudi.pca(sup_var, scale = TRUE, scannf = FALSE)$tab)
-pca$supv <- supv$cosup
+pca$supv <- supcol(pca, dudi.pca(sup_var, scale = TRUE, scannf = FALSE)$tab)
 
 iner <- inertia.dudi(pca, row.inertia = TRUE, col.inertia = TRUE)
 
@@ -31,7 +29,7 @@ test_that("Variables results are equal", {
 })
 
 test_that("Qualitative supplementary variables results are equal", {
-  expect_equal(as.vector(round(pca$supv[,1],3)),
+  expect_equal(as.vector(round(pca$supv$cosup[,1],3)),
                res$vars[res$vars$Type == "Supplementary" & 
                           res$vars$Class == "Quantitative" & res$vars$Axis == "1", "Coord"])
 })
@@ -47,7 +45,7 @@ test_that("Individuals results are equal", {
 })  
   
 test_that("Supplementary individuals results are equal", {
-  expect_equal(as.vector(round(pca$supi[,4],3)),
+  expect_equal(as.vector(round(pca$supi$lisup[,4],3)),
                data.frame(res$ind)[res$ind$Type == "Supplementary" & res$ind$Axis == "4", "Coord"])
 })